requirements.txt
numpy==1.26.4
pandas==2.2.1
statsmodels==0.14.1
matplotlib==3.8.3
seaborn==0.13.2
PyYAML==6.0.1
scipy==1.12.0
